+++ CRM/Xdedupe/Config.php
@@ -120,7 +120,7 @@ class CRM_Xdedupe_Config implements EventSubscriberInterface
         CRM_Xdedupe_Resolver_CustomGroupIntegral::addAllResolvers($xdedupe_list->list);
         CRM_Xdedupe_Resolver_MultiSelect::addAllResolvers($xdedupe_list->list);
         CRM_Xdedupe_Resolver_CustomField::addAllResolvers($xdedupe_list->list);
-        CRM_Xdedupe_Resolver_CustomGroupByField::addAllResolvers($xdedupe_list->list);
+        CRM_Xdedupe_Resolver_CustomGroupMerge::addAllResolvers($xdedupe_list->list);
     }
 
     /**
